Mark identical (untranslated) texts with asterisks
The three asterisks in the `language.h` file mean that the text is the same as in English (as per 1st November 2023). This will help to identify texts which have not been translated so far.

Of course, some texts are the same in many languages, like `TV` or `SMS`, so feel free to remove them when updating your language.