Update language.h
Hello! 

I reworked the Russian language of the receiver menu.

In the old version, some lines did not fit into the screen due to their length and language features. Also, in some places words were abbreviated, which looks unaesthetic and not entirely correct.

I used synonyms and shorter expressions so that the lines fit into the display on the screen and the paragraphs were pleasant to read.

I am a native russian speaker. You might be interested in including my version of the translation in the overall project. 

I will be very glad to be useful. Thank you!